Update from upstream

+++ streamonitor/downloaders/hls.py
@@ -1,41 +1,58 @@
 import m3u8
 import os
-import requests
 import subprocess
 from threading import Thread
 from ffmpy import FFmpeg, FFRuntimeError
 from time import sleep
-from parameters import DEBUG
+from parameters import DEBUG, CONTAINER, SEGMENT_TIME, FFMPEG_PATH
+
+_http_lib = None
+if not _http_lib:
+    try:
+        import pycurl_requests as requests
+        _http_lib = 'pycurl'
+    except ImportError:
+        pass
+if not _http_lib:
+    try:
+        import requests
+        _http_lib = 'requests'
+    except ImportError:
+        pass
+if not _http_lib:
+    raise ImportError("Please install requests or pycurl package to proceed")
 
 
-def getVideoNativeHLS(self, url, filename):
+def getVideoNativeHLS(self, url, filename, m3u_processor=None):
     self.stopDownloadFlag = False
     error = False
-    tmpfilename = filename[:-len('.mp4')] + '.tmp.ts'
-
-    def debug_(message):
-        self.debug(message, filename + '.log')
+    tmpfilename = filename[:-len('.' + CONTAINER)] + '.tmp.ts'
+    session = requests.Session()
 
     def execute():
         nonlocal error
-        downloaded_list = []
+        downloaded_list: set[str] = set()
         with open(tmpfilename, 'wb') as outfile:
             did_download = False
             while not self.stopDownloadFlag:
-                r = requests.get(url, headers=self.headers, cookies=self.cookies)
-                chunklist = m3u8.loads(r.content.decode("utf-8"))
+                r = session.get(url, headers=self.headers, cookies=self.cookies)
+                content = r.content.decode("utf-8")
+                if m3u_processor:
+                    content = m3u_processor(content)
+                chunklist = m3u8.loads(content)
                 if len(chunklist.segments) == 0:
                     return
-                for chunk in chunklist.segments:
-                    if chunk.uri in downloaded_list:
-                        continue
+                all_segments = chunklist.segment_map + chunklist.segments
+                chunks = [chunk.uri for chunk in all_segments if chunk.uri and chunk.uri not in downloaded_list]
+                if not chunks:
+                    sleep(1)
+                for chunk_uri in chunks:
                     did_download = True
-                    downloaded_list.append(chunk.uri)
-                    chunk_uri = chunk.uri
-                    debug_('Downloading ' + chunk_uri)
+                    downloaded_list.add(chunk_uri)
+                    self.debug('Downloading ' + chunk_uri)
                     if not chunk_uri.startswith("https://"):
                         chunk_uri = '/'.join(url.split('.m3u8')[0].split('/')[:-1]) + '/' + chunk_uri
-                    m = requests.get(chunk_uri, headers=self.headers, cookies=self.cookies)
+                    m = session.get(chunk_uri, headers=self.headers, cookies=self.cookies)
                     if m.status_code != 200:
                         return
                     outfile.write(m.content)
@@ -56,11 +73,25 @@ def getVideoNativeHLS(self, url, filename):
     if error:
         return False
 
+    if not os.path.exists(tmpfilename):
+        return False
+
+    if os.path.getsize(tmpfilename) == 0:
+        os.remove(tmpfilename)
+        return False
+
     # Post-processing
     try:
         stdout = open(filename + '.postprocess_stdout.log', 'w+') if DEBUG else subprocess.DEVNULL
         stderr = open(filename + '.postprocess_stderr.log', 'w+') if DEBUG else subprocess.DEVNULL
-        ff = FFmpeg(inputs={tmpfilename: None}, outputs={filename: '-codec copy'})
+        output_str = '-c:a copy -c:v copy'
+        suffix = ''
+        if SEGMENT_TIME is not None:
+            output_str += f' -f segment -reset_timestamps 1 -segment_time {str(SEGMENT_TIME)}'
+            if hasattr(self, 'filename_extra_suffix'):
+                suffix = self.filename_extra_suffix
+            filename = filename[:-len('.' + CONTAINER)] + '_%03d' + suffix + '.' + CONTAINER
+        ff = FFmpeg(executable=FFMPEG_PATH, inputs={tmpfilename: None}, outputs={filename: output_str})
         ff.run(stdout=stdout, stderr=stderr)
         os.remove(tmpfilename)
     except FFRuntimeError as e:
